backend/requirements.txt
django
djangorestframework
markdown
django-filter
Pillow
django-cors-headers
